Refactor and unify Gradle documentation tasks
Replaced and consolidated multiple task implementations into a more unified, modular structure. Introduced `TransformationTask` as the base class, simplifying logic and removing redundant classes like `AbstractStyleSheetTask`, `XIncludeTask`, and `DocbookHtml`. Improved maintainability by moving shared functionality to `AbstractTransformationTask` and aligning naming conventions for consistency.
